реализовал работу рейтинга товара

+++ server/models/Product.js
@@ -6,7 +6,7 @@ const Product = sequelize.define('product', {
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name: {type: DataTypes.STRING, allowNull: false},
     price: {type: DataTypes.INTEGER, allowNull: false},
-    rating: {type: DataTypes.INTEGER, defaultValue: 0},
+    rating: {type: DataTypes.FLOAT, defaultValue: 0},
     img: {type: DataTypes.STRING, allowNull: false}
 })
